+++ MoodIndigo/app/build.gradle
@@ -29,4 +29,9 @@ dependencies {
     implementation 'com.android.support:design:25.4.0'
     implementation 'com.android.support.constraint:constraint-layout:1.0.2'
     implementation 'com.android.support:support-v4:25.4.0'
+
+    //OWAIS : support design library
+    compile 'com.android.support:appcompat-v7:25.4.0'
+    compile 'com.android.support:design:25.4.0'
+    // Owais
 }
