+++ MoodIndigo/app/src/main/java/com/example/android/moodindigo/Fragments/MainFragment.java
@@ -19,9 +19,11 @@ import com.example.android.moodindigo.R;
 
 public class MainFragment extends Fragment {
 
-    View view;
-    ViewPager viewPager;
-    ViewPagerAdapter viewPagerAdapter;
+    //Mrunmayi
+//    View view;
+//    ViewPager viewPager;
+//    ViewPagerAdapter viewPagerAdapter;
+    //Mrunmayi
     public MainFragment() {
         // Required empty public constructor
     }
@@ -31,66 +33,70 @@ public class MainFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         container.removeAllViews();
+        return inflater.inflate(R.layout.fragment_main, container,false);
 
-        view=inflater.inflate(R.layout.fragment_main, container, false);
-        TabLayout tabLayout = (TabLayout) view.findViewById(R.id.tabs);
-        tabLayout.addTab(tabLayout.newTab().setText("News"));
-        tabLayout.addTab(tabLayout.newTab().setText("Competitions"));
+        //Mrunmayi
+//        view=inflater.inflate(R.layout.fragment_main, container, false);
+//        TabLayout tabLayout = (TabLayout) view.findViewById(R.id.tabs);
+//        tabLayout.addTab(tabLayout.newTab().setText("News"));
+//        tabLayout.addTab(tabLayout.newTab().setText("Competitions"));
+//
+//        viewPager = (ViewPager) view.findViewById(R.id.container);
+//        viewPager.setAdapter(new MainFragment.ViewPagerAdapter(getFragmentManager(), tabLayout.getTabCount()));
+//        viewPager.addOnPageChangeListener(new TabLayout.TabLayoutOnPageChangeListener(tabLayout));
+//        tabLayout.addOnTabSelectedListener(new TabLayout.OnTabSelectedListener() {
+//            @Override
+//            public void onTabSelected(TabLayout.Tab tab) {
+//                viewPager.setCurrentItem(tab.getPosition());
+//            }
+//
+//            @Override
+//            public void onTabUnselected(TabLayout.Tab tab) {
+//
+//            }
+//
+//            @Override
+//            public void onTabReselected(TabLayout.Tab tab) {
+//
+//            }
+//        });
+//        return view;
+//    }
+//
+//    public class ViewPagerAdapter extends FragmentStatePagerAdapter {
+//        int mNumOfTabs;
+//
+//        public ViewPagerAdapter(FragmentManager fm, int NumOfTabs) {
+//            super(fm);
+//            this.mNumOfTabs = NumOfTabs;
+//        }
+//
+//        @Override
+//        public Fragment getItem(int position) {
+//
+//            switch (position) {
+//                case 0:
+//                    NewsFragment tab1 = new NewsFragment();
+//                    return tab1;
+//                case 1:
+//                    CompiFragment tab2 = new CompiFragment();
+//                    return tab2;
+//                default:
+//                    return null;
+//            }
+//        }
+//
+//        @Override
+//        public int getCount() {
+//            return mNumOfTabs;
+//        }
+//    }
 
-        viewPager = (ViewPager) view.findViewById(R.id.container);
-        viewPager.setAdapter(new MainFragment.ViewPagerAdapter(getFragmentManager(), tabLayout.getTabCount()));
-        viewPager.addOnPageChangeListener(new TabLayout.TabLayoutOnPageChangeListener(tabLayout));
-        tabLayout.addOnTabSelectedListener(new TabLayout.OnTabSelectedListener() {
-            @Override
-            public void onTabSelected(TabLayout.Tab tab) {
-                viewPager.setCurrentItem(tab.getPosition());
-            }
+        //Mrunmayi
 
-            @Override
-            public void onTabUnselected(TabLayout.Tab tab) {
 
-            }
 
-            @Override
-            public void onTabReselected(TabLayout.Tab tab) {
-
-            }
-        });
-        return view;
     }
-
-    public class ViewPagerAdapter extends FragmentStatePagerAdapter {
-        int mNumOfTabs;
-
-        public ViewPagerAdapter(FragmentManager fm, int NumOfTabs) {
-            super(fm);
-            this.mNumOfTabs = NumOfTabs;
-        }
-
-        @Override
-        public Fragment getItem(int position) {
-
-            switch (position) {
-                case 0:
-                    NewsFragment tab1 = new NewsFragment();
-                    return tab1;
-                case 1:
-                    CompiFragment tab2 = new CompiFragment();
-                    return tab2;
-                default:
-                    return null;
-            }
-        }
-
-        @Override
-        public int getCount() {
-            return mNumOfTabs;
-        }
-    }
-
-
-
-
 }
